fix: fix sessions-idls package cjs build (#20)
Previously, the cjs build of this package was not a valid package as it did not
have any entry point being shipped and the file `./dist/cjs/index.js` referenced
by the `package.json` did not actually ever get built.

This PR restructures things to explicitly have a separate cjs compatible entry
point so that the built cjs package is actually valid.