fix: ensure styling is consistent and handle swr mutations
This commit attempts to add some reset & font css to ensure that the widget gets
a consistent style, regardless of the context of the surrounding app.  It won't
be perfectly isolated, but it should do a fairly good job in most situations.

This commit also adds some code to bust the token account data cache after
transactions are sent successfully.