fix: only check session if a new wallet address connects (#77)
Previously, any time the provider re-rendered and the wallet entered into a
connected state while it was already connected, we would assume that meant a new
wallet connected, and would enter into the `CheckingStoredSession` state.

This works as long as the provider is in a part of the react tree that doesn't
have unnecessary rerenders.  But if the session provider is used in a context
that causes rerenders unnecessarily, such as in a router that renders the layout
on each page change, then it would cause us to think a new wallet is being
connected on each page change.

This PR just tightens the conditions for assuming a new wallet is connected and
only goes into `CheckingStoredSession` if rerendering with the wallet in a
connected state while already connected AND the wallet address doesn't match the
prior one.

This PR also fixes some minor issues with ensuring that errors and failures when
creating sessions fully reset the session rather than just disconnecting the
wallet.