fix: fix style issues
This PR removes tailwind from the example app; tailwindw was a problem because
tailwind sets some global styles, which meant that some of the styles we were
shipping only worked in tailwind and weren't discovered until after releasing a
version and trying it outside of tailwind.  This has caused numerous style bugs.

This PR also fixes a few style bugs in the sdk (most of which were previously
hidden due to tailwind):

- Set `color-scheme` on the reset element to ensure the containing app's color
scheme doesn't affect any sessions widgets
- Add a space between the token amount and token symbol in the wallet widget
- Make sure the "Get tokens" label has a color set