fix: replace react-timeago with a custom component
The `react-timeago` library does not ship esm correctly and it's caused issues
for the typing, and some of the developers using the sdk have reported that they
can't use the latest release due to issues with their environment interacting
with this library.

It's a pretty simple component.  Rather than fighting with the library to get it
to work correctly, I figure it's better to just home-roll this.