fix: use custom groupBy
Apparently `Object.groupBy` still is not as widely supported as we think, and is
in particular not supported on iPhone 14 on any browser.  We don't really have a
lot of polyfills that we usually need and this isn't a hot path, so rather than
worrying with a proper polyfill, this commit just adds a manual `groupBy`
implementation.