feat: add mainnet support
This commit adds support for mainnet to the Fogo Sessions typescript sdks.

Implementing this turned into a somewhat larger refactor.

It became obvious while working on this that it did not make sense to continue
using `ConnectionProvider` from `@solana/wallet-adapter-react`, and that the
information for the default RPC url should be part of the typescript SDK, not
the React SDK.

However, `createSessionContext` is an async function because it must resolve the
chain ID, address lookup tables, and sponsor.  So if we used this to create the
RPC / web3.js `Connection`, then those would be unavailable until the session
context was finished resolving.

Because there are use cases where we need an rpc before the session context is
done initializing, it makes sense to separate out the asynchronous parts of
context from the synchronous.

This commit thus introduces a new function `createSessionConnection` which is
synchronous, and exposes the `Connection` & rpc via the session contexts.