feat: add callback to open session panel on bridge in page (#271)
We want some apps to have a way to have a button on the page which will open the
session widget directly onto the page to bridge USDC in from Solana.

This is a bit of a hacky way of doing it, but it's the least-invasive way to add
the functionality with a decent client API.  I think this is an OK tradeoff
since I doubt this feature will be kept around long after the initial mainnet
launch.