fix: add network as a cache key for cached data (#307)
Before this change, we did not have the network as a cache key for data cached
with `useData`, or for the session itself.  As a result, when building an app
with a selector to change networks, when changing networks:

1. Stale data would be shown in the widget for the wallet on the prior network
2. The session key would be erased on each network switch as the key would be
   seen as invalid on the new network

As of this change, we now segregate the caches between mainnet & testnet, so
switching between networks is much smoother.

We also modify the init hook so that we re-check the cached connection when the
network changes, rather than only on app startup.