feat: show native balance; add enable native transfers w/ wallet
This change shows native token balance in the session widget and adds a stopgap
mechanism for enabling native transfers, by allowing transactions to be signed
with the wallet instead of the session.

The UX here has some downsides:

1. The wallet will show a big scary warning message about the transaction being
   untrusted
2. The wallet will show SOL instead of FOGO as the token being transferred
3. Currently I haven't added the necessary changes to enable sending tokens
   using this new mechanism (but this is something I can do in a follow-up if we
   decide to move forward on this)

However, despite these downsides, this will allow users to interact with native
tokens even before sessions can actually support them, so we might consider this
as a temporary stopgap solution.