chore: simplify IDL generation (#370)
I'm fairly sure the root cause of the build failures we're seeing in Vercel is
because IDL generation cannot be done in parallel; if it is done in parallel the
Rust builds write to the same build cache and that causes race conditions.

This is avoided locally because Rust will wait for a lock on the build
directory, but for some reason that same wait isn't happening in Vercel and the
idl generation is happening in parallel, triggering the race condition.

I was going to look into why the build directory lock wait isn't happening in
Vercel, but honestly, having the turbo tasks broken down for the idl generation
like this really only made sense if we intended to run the idl generation in
parallel.  If we don't plan to do that, we might as well clean this up and
simplify it considerably and generate the idls in sequence instead.