Component library storybook (#428)
* chore: delete old component library

* feat: set up component library

This commit initializes the component library package.  So that we don't have to
publish it or have it as a separate package dependency of the react sdk, it is
symlinked to the react sdk src and is distributed along with the react sdk code.
With this approach, components that are in the component library but are not in
use by the react sdk will be included in the built package, however they won't
be accessible to consumers due to subpath exports.

* lockfile

* add storybook

* lockfile

* revert

* format

* ignore stories

* add next

* add vercel.json

---------

Co-authored-by: Connor Prussin <connor@prussin.net>