fix: fix build (#431)
This commit adds the `build:package-json` script so the component library dist
is indeed transformed into a fully correct build artifact.  In order to ensure
that the component library does not get published, I added a new option
`--no-remove-private` to `transform-package-json` -- this way, for this package,
the dist artifact continues to be marked as private, and the package is not
deployed.